Explained why I forked this repository in README.
In 2021, I began creating dedicated /Documentation branches
in the repositories I have forked in Github, in order to better
document why I forked that repository, and what work I had done
or intended to do in it.  This commit should be the only difference
from main/master in this branch.